[SWWEB-2284]: New User Registration (#1156)
* [SWWEB-2284]: front-end

* [SWWEB-2284]: front-end correct success/error message handling

* [SWWEB-2284]: back-end

* [SWWEB-2284]: back-end fixes + approvers email

* [SWWEB-2284]: fetching email addresses from maximo

* [SWWEB-2284]: send notification email to user

* [SWWEB-2284]: userprofile.deletable property for delete validation

* [SWWEB-2284]: activation email + approval instructions

* [SWWEB-2284]: custom usersetup error page

+++ softWrench.sW4/Security/Services/UserProfileManager.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using cts.commons.persistence;
@@ -137,6 +138,9 @@ namespace softWrench.sW4.Security.Services {
                     var profile = FindById(profileId);
                     if (profile == null) {
                         return;
+                    }
+                    if (!profile.Deletable) {
+                        throw new InvalidOperationException(string.Format("The userprofile '{0}' is not deletable.", profile.Name));
                     }
                     _dao.DeleteCollection(profile.ApplicationPermissions);
                     _dao.ExecuteSql("delete from sw_user_userprofile where profile_id = {0}".Fmt(profile.Id));
